requirements.txt
streamlit
PyPDF2
python-docx
transformers
googletrans
